requirements.txt
aiosqlite==0.21.0
altgraph==0.17.2
anyio==4.3.0
asgiref==3.6.0
attrs==20.3.0
backports.entry-points-selectable==1.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.8
cryptography==41.0.3
defusedxml==0.8.0rc2
distlib==0.3.6
distro==1.9.0
dj-rest-auth==7.0.1
Django==4.2
django-cors-headers==4.3.1
django-countries==7.6.1
django-debug-toolbar==4.2.0
django-environ==0.12.0
django-filter==23.3
django-phonenumber-field==7.3.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
drf-nested-routers==0.93.4
exceptiongroup==1.1.2
filelock==3.12.0
frappe-bench==5.23.0
future==0.18.2
gitdb==4.0.12
GitPython==3.1.44
h11==0.14.0
honcho==2.0.0
httpcore==1.0.5
httpx==0.27.0
idna==3.4
iniconfig==2.0.0
instagrapi==1.17.10
jellyfish==0.11.2
Jinja2==3.1.5
lib50==3.0.7
macholib==1.15.2
markdown2==2.4.10
MarkupSafe==3.0.2
mysqlclient==2.1.1
numpy==1.25.0
oauthlib==3.2.2
openai==1.20.0
packaging==23.1
pandas==2.0.3
pexpect==4.8.0
phonenumbers==8.13.35
Pillow==10.0.0
pipenv==2023.4.20
platformdirs==3.4.0
plotly==5.24.1
pluggy==1.2.0
ptyprocess==0.7.0
pycparser==2.21
pycryptodomex==3.17
pydantic==1.10.7
PyJWT==2.8.0
PySocks==1.7.1
pytest==7.4.0
python-crontab==2.6.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytube==15.0.0
pytz==2023.3
PyYAML==6.0
Random-Word==1.0.11
requests==2.31.0
requests-oauthlib==2.0.0
semantic-version==2.10.0
six==1.15.0
smmap==5.0.2
sniffio==1.3.1
social-auth-app-django==5.4.2
social-auth-core==4.5.4
SQLAlchemy==2.0.17
sqlparse==0.4.3
submit50==3.1.1
tenacity==9.0.0
termcolor==1.1.0
tomli==2.0.1
tqdm==4.66.2
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
virtualenv==20.22.0
virtualenv-clone==0.5.7
celery>=5.3
pika
